+++ frameworks/runtime-src/Classes/curllua/Lua-utility.c
@@ -32,7 +32,7 @@ const char* luaL_getstrfield(lua_State* L, const char* key) {
   return val;
 }
 
-const char* luaL_getlstrfield(lua_State* L, const char* key, int *len) {
+const char* luaL_getlstrfield(lua_State* L, const char* key, size_t *len) {
   const char *val;
   lua_getfield(L, -1, key);
   val = lua_isnil(L, -1)? NULL : lua_tolstring(L, -1, len);
